manifest: persist SSTable metadata and ensure uint64 IDs - Replace SSTablePaths with SSTables []SSTableMetadata - Recover SSTables with IDs - Persist IDs during flush - Persist IDs during compaction - Update ensureSafeNextSSTableID to use table ID directly - Update manifest tests to use SSTableMetadata - Add DB restart/compaction tests - Use uint64 for SSTable, WAL IDs

+++ internal/memtable/memtable.go
@@ -11,7 +11,7 @@ type MemTable struct {
 
 type ImmutableMemTable struct {
 	MemTable *MemTable
-	WalID    int
+	WalID    uint64
 }
 
 func New() *MemTable {
